Modifications to compile and run assimp with the PlayStation3 homebrew toolchain.
Look for the __lv2ppu__ symbol to be defined, and do the appropriate thing.

Certain importers supported by assimp have global data initialized
statically. Including these importers in the assimp build causes
programs linked to the library to hang on startup; removing them
allows the program to run. I haven't tried to see exactly which
importers cause the problem, or modified their implementation so that
static initialization isn't performed. Might do that in the future,
though.

assimp uses the realpath function, which isn't provided by the
toolchain's libc (I'm not sure if the PS3's filesystem even supports
symbolic links). Replace the call to realpath() with a strncpy().